int -> int64_t in in Typed and PropertyValue
Reviewers: buda, mferencevic

Reviewed By: buda, mferencevic

Subscribers: pullbot

Differential Revision: https://phabricator.memgraph.io/D106

+++ src/communication/bolt/v1/serialization/bolt_serializer.hpp
@@ -119,7 +119,7 @@ class BoltSerializer {
         encoder.write_string(value.Value<std::string>());
         return;
       case PropertyValue::Type::Int:
-        encoder.write_integer(value.Value<int>());
+        encoder.write_integer(value.Value<int64_t>());
         return;
       case PropertyValue::Type::Double:
         encoder.write_double(value.Value<double>());
